% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_attributes.R
\name{aggregate_by_attributes}
\alias{aggregate_by_attributes}
\title{Aggregate PSMs to Protein Level}
\usage{
aggregate_by_attributes(
  PSM,
  col_of_reporter_ion_int,
  ranger = NULL,
  pred_error = NULL,
  ratio_calc = F
)
}
\arguments{
\item{PSM}{A data frame containing all PSMs to be aggregated.}

\item{col_of_reporter_ion_int}{A vector of column names representing reporter ion intensities across different channels.}

\item{ranger}{The random forest model to be applied for aggregation.}

\item{pred_error}{The predicted level of inaccuracy for the PSMs, obtained from external sources. Either the \code{ranger} model or \code{pred_error} must be specified.}

\item{ratio_calc}{A logical value indicating whether relative reporter intensities are calculated using the total reporter intensities across all channels.}
}
\value{
A data frame containing protein abundance estimates.
}
\description{
Aggregates PSMs using a random forest model.
}
