% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_benchmark_set_2.R
\docType{data}
\name{benchmark.set.2}
\alias{benchmark.set.2}
\title{Benchmark Set 2 from FragPipe}
\format{
A data frame with 19566 rows and 21 variables:
\describe{
\item{Peptide}{Peptide amino acid sequence}
\item{Charge}{The charge state of the identified peptide}
\item{Calibrated Observed Mass}{Mass of the identified peptide after m/z calibration (in Da)}
\item{Calibrated Observed M/Z}{Mass-to-charge ratio of the peptide ion after m/z calibration}
\item{Delta Mass}{Difference between calibrated observed peptide mass and theoretical peptide mass (in Da)}
\item{Hyperscore}{The similarity score between observed and theoretical spectra}
\item{Number of Missed Cleavages}{Number of potential enzymatic cleavage sites within the identified sequence}
\item{Intensity}{Raw integrated precursor abundance for each PSM}
\item{Assigned Modifications}{Post-translational modifications within the identified sequence}
\item{Purity}{The proportion of total ion abundance in the inclusion window from the precursor}
\item{Protein}{Protein sequence header corresponding to the identified peptide sequence}
\item{H1+Y1_1, H1+Y4_1, H1+Y10_1, H1+Y1_2, H1+Y4_2, H1+Y10_2, H1+Y1_3, H1+Y4_3, H1+Y10_3}{Reporter ion intensities for various samples. For example, "H1+Y4_1" represents the reporter ion intensity for the first replicate (_1) of the group H1+Y4, which contains a constant level of human proteins (H1) and a spiking of yeast proteins (Y4) at four times the base level}
\item{Replicate}{The mass spectrometry run where the PSM is detected}
}
}
\usage{
benchmark.set.2
}
\description{
This data frame represents peptide spectrum matches (PSMs) in Benchmark Set 2, obtained from the search results of FragPipe. Columns unnecessary for the AWAggregator have been removed from the sample data.
}
\keyword{datasets}
