% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_no_CV_2.R
\docType{data}
\name{child.nodeIDs.no.CV}
\alias{child.nodeIDs.no.CV}
\title{Pre-trained Random Forest Model Without Coefficient of Variation (Part 2)}
\format{
A pre-trained random forest model that excludes average CV as a feature (Part 2).
}
\usage{
child.nodeIDs.no.CV
}
\description{
This object represents the second part of a pre-trained random forest model that does not include the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
