% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_2.R
\docType{data}
\name{child.nodeIDs}
\alias{child.nodeIDs}
\title{Pre-trained Random Forest Model Using Coefficient of Variation (Part 2)}
\format{
A pre-trained random forest model that utilizes average CV as a feature (Part 2).
}
\usage{
child.nodeIDs
}
\description{
This object represents the second part of a pre-trained random forest model that incorporates the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
