% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_PD_format.R
\name{convert_PD_format}
\alias{convert_PD_format}
\title{Convert Proteome Discoverer Output to Required Input Format}
\usage{
convert_PD_format(PSM, protein, col_of_reporter_ion_int)
}
\arguments{
\item{PSM}{A data frame containing the PSM table from Proteome Discoverer to be converted.}

\item{protein}{A data frame containing the corresponding protein table from Proteome Discoverer.}

\item{col_of_reporter_ion_int}{A vector of column names for reporter ion intensities across different channels.}
}
\value{
A data frame in the format required by AWAggregator.
}
\description{
Converts output from Proteome Discoverer into the input format required by AWAggregator.
}
