% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a Random Forest Model}
\usage{
fit_model(
  PSM,
  num.trees = 500,
  use_average_CV = T,
  importance = F,
  seed,
  applied_attributes = c(NA)
)
}
\arguments{
\item{PSM}{A data frame containing all PSMs used for training.}

\item{num.trees}{The number of trees to include in the random forest model.}

\item{use_average_CV}{A logical value indicating whether to include the average CV attribute in the training. This parameter is ignored if \code{applied_attributes} is specified.}

\item{importance}{A logical value indicating whether to assess the importance of attributes.}

\item{seed}{An integer seed for random number generation in the random forest.}

\item{applied_attributes}{A vector of attribute names to be used in training, replacing the default attributes.}
}
\value{
A trained random forest model.
}
\description{
Trains a random forest model to predict the level of quantitative inaccuracy of PSMs.
}
