% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit a Random Forest Model}
\usage{
fitModel(
  PSM,
  numTrees = 500,
  useAvgCV = TRUE,
  importance = FALSE,
  seed,
  appliedAttributes = c(NA)
)
}
\arguments{
\item{PSM}{A data frame containing all PSMs used for training.}

\item{numTrees}{The number of trees to include in the random forest model.}

\item{useAvgCV}{A logical value indicating whether to include the average CV attribute in the training. This parameter is ignored if \code{appliedAttributes} is specified.}

\item{importance}{A logical value indicating whether to assess the importance of attributes.}

\item{seed}{An integer seed for random number generation in the random forest.}

\item{appliedAttributes}{A vector of attribute names to be used in training, replacing the default attributes.}
}
\value{
A trained random forest model.
}
\description{
Trains a random forest model to predict the level of quantitative inaccuracy of PSMs.
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub()
benchmarkSet3 <- eh[['EH9639']]
samples <- c('H1+Y0_1', 'H1+Y1_1', 'H1+Y5_1',
    'H1+Y10_1', 'H1+Y10_2', 'H1+Y5_2', 'H1+Y1_2', 'H1+Y0_2')
groups <- c('H1+Y0', 'H1+Y1', 'H1+Y5',
    'H1+Y10', 'H1+Y10', 'H1+Y5', 'H1+Y1', 'H1+Y0')
PSM <- getAttributes(
    PSM=benchmarkSet3,
    # TMTpro tag (304.2071) and N-ethylmaleimide (125.0476) are applied as fixed PTMs
    fixedPTM=c('304.2071', '125.0476'),
    colOfReporterIonInt=samples,
    groups=groups
)
PSM <- getAvgScaledErrorOfLog2FC(
    PSM=PSM,
    colOfReporterIonInt=samples,
    groups=groups,
    expectedRelativeAbundance=list(`H1+Y0`=0, `H1+Y1`=1, `H1+Y5`=5, `H1+Y10`=10),
    speciesAtConstLevel='HUMAN'
)
regr <- fitModel(PSM, useAvgCV=TRUE, seed=3979)
}
