% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes.R
\name{get_attributes}
\alias{get_attributes}
\title{Get attributes for PSMs}
\usage{
get_attributes(
  PSM,
  fixed_PTMs,
  col_of_reporter_ion_int,
  groups,
  groups_excluded_from_CV = NA,
  set_progress_bar = T
)
}
\arguments{
\item{PSM}{A data frame containing all PSMs used for training.}

\item{fixed_PTMs}{A numeric vector with the masses of fixed post-translational modifications (PTMs) in Da. Other PTMs will be treated as variable PTMs.}

\item{col_of_reporter_ion_int}{A vector of column names for reporter ion intensities across different channels.}

\item{groups}{A vector specifying sample groups.}

\item{groups_excluded_from_CV}{A vector of sample groups excluded from average CV calculations, which may occur due to a zero spike-in concentration for a species.}

\item{set_progress_bar}{A logical value indicating whether to display a progress bar.}
}
\value{
A data frame containing the PSM table with attributes required for the random forest model.
}
\description{
Retrieves attributes required for training or test sets.
}
