% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_scaled_error_of_log2FC.R
\name{get_avg_scaled_error_of_log2FC}
\alias{get_avg_scaled_error_of_log2FC}
\title{Get Average Scaled Error of log2FC for PSMs in a Spike-in Dataset}
\usage{
get_avg_scaled_error_of_log2FC(
  PSM,
  col_of_reporter_ion_int,
  groups,
  expected_relative_abundance,
  species_at_const_level
)
}
\arguments{
\item{PSM}{A data frame containing all PSMs used for training.}

\item{col_of_reporter_ion_int}{A vector of column names for reporter ion intensities across different channels.}

\item{groups}{A vector specifying sample groups.}

\item{expected_relative_abundance}{A named list where group names are keys and the corresponding expected relative abundance values for species at varying concentrations are provided as values. Unknown ratios can be designated as NA.}

\item{species_at_const_level}{A string specifying the species that are spiked in at a constant level.}
}
\value{
A data frame containing PSMs with Average Scaled Error of log2FC values required for the random forest model.
}
\description{
Calculates the Average Scaled Error of log2FC values required for training sets.
}
