% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_metric.R
\name{get_dist_metric}
\alias{get_dist_metric}
\title{Get Distance Metric}
\usage{
get_dist_metric(PSM, channel, set_progress_bar = T)
}
\arguments{
\item{PSM}{A data frame containing the PSMs for which distance metrics are to be calculated.}

\item{channel}{A vector specifying the channels used for calculating the distance metric.}

\item{set_progress_bar}{A logical value indicating whether to display a progress bar.}
}
\value{
A vector of distance metrics for the specified PSMs.
}
\description{
Calculates the distance metric for PSMs.
}
