% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttributes.R
\name{getAttributes}
\alias{getAttributes}
\title{Get attributes for PSMs}
\usage{
getAttributes(
  PSM,
  fixedPTMs,
  colOfReporterIonInt,
  groups,
  groupsExcludedFromCV = NA,
  setProgressBar = TRUE
)
}
\arguments{
\item{PSM}{A data frame containing all PSMs used for training.}

\item{fixedPTMs}{A numeric vector with the masses of fixed post-translational modifications (PTMs) in Da. Other PTMs will be treated as variable PTMs.}

\item{colOfReporterIonInt}{A vector of column names for reporter ion intensities across different channels.}

\item{groups}{A vector specifying sample groups.}

\item{groupsExcludedFromCV}{A vector of sample groups excluded from average CV calculations, which may occur due to a zero spike-in concentration for a species.}

\item{setProgressBar}{A logical value indicating whether to display a progress bar.}
}
\value{
A data frame containing the PSM table with attributes required for the random forest model.
}
\description{
Retrieves attributes required for training or test sets.
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub()
benchmarkSet3 <- eh[['EH9639']]
samples <- c('H1+Y0_1', 'H1+Y1_1', 'H1+Y5_1',
    'H1+Y10_1', 'H1+Y10_2', 'H1+Y5_2', 'H1+Y1_2', 'H1+Y0_2')
groups <- c('H1+Y0', 'H1+Y1', 'H1+Y5',
    'H1+Y10', 'H1+Y10', 'H1+Y5', 'H1+Y1', 'H1+Y0')
PSM <- getAttributes(
    PSM=benchmarkSet3,
    fixedPTM=c('304.2071', '125.0476'),
    colOfReporterIonInt=samples,
    groups=groups,
    groupsExcludedFromCV='H1+Y0'
)
}
