% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistMetric.R
\name{getDistMetric}
\alias{getDistMetric}
\title{Get Distance Metric}
\usage{
getDistMetric(PSM, channel, setProgressBar = TRUE)
}
\arguments{
\item{PSM}{A data frame containing the PSMs for which distance metrics are to be calculated.}

\item{channel}{A vector specifying the channels used for calculating the distance metric.}

\item{setProgressBar}{A logical value indicating whether to display a progress bar.}
}
\value{
A vector of distance metrics for the specified PSMs.
}
\description{
Calculates the distance metric for PSMs.
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub()
benchmarkSet3 <- eh[['EH9639']]
samples <- c('H1+Y0_1', 'H1+Y1_1', 'H1+Y5_1',
    'H1+Y10_1', 'H1+Y10_2', 'H1+Y5_2', 'H1+Y1_2', 'H1+Y0_2')
df <- getDistMetric(
    PSM=benchmarkSet3,
    channel=samples,
    setProgressBar=TRUE
)
}
