% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pretrained_model.R
\name{load_pretrained_model}
\alias{load_pretrained_model}
\title{Load a Pre-trained Random Forest Model}
\usage{
load_pretrained_model(use_average_CV = T)
}
\arguments{
\item{use_average_CV}{A logical value indicating whether the average CV attribute was used in the training.}
}
\value{
The pre-trained random forest model.
}
\description{
Loads a pre-trained random forest model for predicting the level of quantitative inaccuracy of PSMs.
}
