% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_training_sets.R
\name{merge_training_sets}
\alias{merge_training_sets}
\title{Merge Training Sets}
\usage{
merge_training_sets(PSM_list, num_PSMs, set_progress_bar = T, seed)
}
\arguments{
\item{PSM_list}{A named list where dataset names are keys and the corresponding data frames of PSMs used for training are values.}

\item{num_PSMs}{The minimum number of PSMs to extract from each dataset for merging.}

\item{set_progress_bar}{A logical value indicating whether to display a progress bar.}

\item{seed}{An integer seed for random selection of PSMs.}
}
\value{
A data frame containing the merged PSM table from a subset of each input dataset.
}
\description{
Extracts a similar number of PSMs from each input dataset and merges them into a single training set.
}
