% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_no_CV_1.R
\docType{data}
\name{regr.no.CV}
\alias{regr.no.CV}
\title{Pre-trained Random Forest Model Without Coefficient of Variation (Part 1)}
\format{
A pre-trained random forest model that excludes average CV as a feature (Part 1).
}
\usage{
regr.no.CV
}
\description{
This object represents the first part of a pre-trained random forest model that does not include the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
