% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_1.R
\docType{data}
\name{regr}
\alias{regr}
\title{Pre-trained Random Forest Model Using Coefficient of Variation (Part 1)}
\format{
A pre-trained random forest model that utilizes average CV as a feature (Part 1).
}
\usage{
regr
}
\description{
This object represents the first part of a pre-trained random forest model that incorporates the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
