% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample_prot_PD.R
\docType{data}
\name{sample.prot.PD}
\alias{sample.prot.PD}
\title{Sample Protein Data from Protein Discoverer}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{Accession}{The unique identifier assigned to the protein}
\item{Description}{The name of the protein exclusive of the identifier
that appears in the Accession column}
}
}
\usage{
data(sample.prot.PD)
}
\description{
This data frame represents sample proteins A0AV96, A0AVF1, A0AVT1, A0FGR8,
and A0M8Q6, obtained from the search results of Proteome Discoverer. Columns
unnecessary for the AWAggregator have been removed from the sample data.
}
\keyword{datasets}
