% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_3.R
\docType{data}
\name{split.values.1}
\alias{split.values.1}
\title{Pre-trained Random Forest Model Using Coefficient of Variation (Part 3)}
\format{
A pre-trained random forest model that utilizes average CV as a feature (Part 3).
}
\usage{
split.values.1
}
\description{
This object represents the third part of a pre-trained random forest model that incorporates the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
