% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_4.R
\docType{data}
\name{split.values.2}
\alias{split.values.2}
\title{Pre-trained Random Forest Model Using Coefficient of Variation (Part 4)}
\format{
A pre-trained random forest model that utilizes average CV as a feature (Part 4).
}
\usage{
split.values.2
}
\description{
This object represents the fourth part of a pre-trained random forest model that incorporates the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
