% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_no_CV_3.R
\docType{data}
\name{split.values.no.CV.1}
\alias{split.values.no.CV.1}
\title{Pre-trained Random Forest Model Without Coefficient of Variation (Part 3)}
\format{
A pre-trained random forest model that excludes average CV as a feature (Part 3).
}
\usage{
split.values.no.CV.1
}
\description{
This object represents the third part of a pre-trained random forest model that does not include the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
