% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_no_CV_4.R
\docType{data}
\name{split.values.no.CV.2}
\alias{split.values.no.CV.2}
\title{Pre-trained Random Forest Model Without Coefficient of Variation (Part 4)}
\format{
A pre-trained random forest model that excludes average CV as a feature (Part 4).
}
\usage{
split.values.no.CV.2
}
\description{
This object represents the fourth part of a pre-trained random forest model that does not include the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
