% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_no_CV_5.R
\docType{data}
\name{split.varIDs.no.CV}
\alias{split.varIDs.no.CV}
\title{Pre-trained Random Forest Model Without Coefficient of Variation (Part 5)}
\format{
A pre-trained random forest model that excludes average CV as a feature (Part 5).
}
\usage{
split.varIDs.no.CV
}
\description{
This object represents the fifth part of a pre-trained random forest model that does not include the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
