% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regr_5.R
\docType{data}
\name{split.varIDs}
\alias{split.varIDs}
\title{Pre-trained Random Forest Model Using Coefficient of Variation (Part 5)}
\format{
A pre-trained random forest model that utilizes average CV as a feature (Part 5).
}
\usage{
split.varIDs
}
\description{
This object represents the fifth part of a pre-trained random forest model that incorporates the average coefficient of variation (CV) as a feature.
}
\keyword{datasets}
